\name{COAP-package}
\alias{COAP-package}
\alias{COAP}
\docType{package}
\title{
\packageTitle{COAP}
}
\description{
\packageDescription{COAP}
}
\details{
The current Poisson factor models often assume that the factors are unknown, which overlooks the explanatory potential of certain observable covariates. This study focuses on high dimensional settings, where the number of the count response variables and/or covariates can diverge as the sample size increases. A covariate-augmented overdispersed Poisson factor model is proposed to jointly perform a high-dimensional Poisson factor analysis and estimate a large coefficient matrix for overdispersed count data. 
}
\author{
\packageAuthor{COAP}

Maintainer: \packageMaintainer{COAP}
}
\references{
  None
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<GFM>}} ~~
}
