\name{COAP-package}
\alias{COAP-package}
\alias{COAP}
\docType{package}
\title{
\packageTitle{COAP}
}
\description{
\packageDescription{COAP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{COAP}
\packageIndices{COAP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{COAP}

Maintainer: \packageMaintainer{COAP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
